requirements.txt
DO NOT TRY TO INSTALL THIS PYTHON APPLICATION - CTF Purpose
Flask==2.3.2
flask-cors==4.0.0
gunicorn==20.1.0
numpy==1.24.3
pandas==2.0.2
tensorflow==2.13.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.3
google-cloud-secret-manager==2.16.3
google-cloud-pubsub==2.18.1
google-cloud-logging==3.5.0
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
prometheus-client==0.17.1
redis==4.6.0
celery==5.3.1